requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
pillow>=10.0.0
pyarrow>=14.0.0
tqdm>=4.66.0
wandb>=0.16.0
numpy>=1.24.0
pyyaml>=6.0
accelerate>=0.25.0